refactor(rpc/git): split the git dispatcher into a module directory
rpc/git.rs (1539 lines: six responsibilities in one impl plus 900 lines
of tests) becomes rpc/git/:

  targets         which checkout an op acts on (Target, target(), with_repo)
  publishing      open_pr, git_push, branch materialization, approval gate
  commands        git_status, git_fetch, run_git_command with the config guard
  review_threads  pr_threads, reply_thread, resolve_thread
  args            arg readers and the option-like ref refusal
  mod.rs          GitDispatcher, builders, dispatch routing, signals, EVENT_*

Methods spread across files as separate impl blocks; pub(super) only
where a sibling calls in. The git_status arm becomes a method in
commands so mod.rs routes only. Tests move to rpc/git/tests/ split the
same way, with shared fixtures in tests/support.rs. rpc.rs repoints its
#[path] at the new mod.rs.

Frozen: GitDispatcher and its builders, EVENT_* names and payloads, op
and error strings, gate ordering in open_pr and git_push, tests.

cargo test 1628 passed; fmt and clippy -D warnings clean.
